+++ src/shopifyextractor.nim
@@ -1,5 +1,5 @@
 import std/asyncdispatch
-from std/uri import parseUri
+from std/uri import parseUri, `$`
 from std/htmlparser import parseHtml
 from std/strutils import contains
 # from std/xmltree import XmlNode, innerText, xnElement
@@ -10,6 +10,8 @@ import pkg/nimquery
 import pkg/unifetch
 
 import shopifyextractor/types
+import shopifyextractor/product
+export product
 
 proc fetchSitemap(url: Uri or string): Future[XmlNode] {.async.} =
   let
@@ -18,21 +20,17 @@ proc fetchSitemap(url: Uri or string): Future[XmlNode] {.async.} =
   close uni
   result = parseHtml resp.body
 
-proc extractShopify(url: string): Future[ShopifyEcommerce] {.async.} =
-  ## Extracts all products of a Shopify ecommerce
+proc getAllProducts(url: string): Future[seq[string]] {.async.} =
+  ## Returns all products of a Shopify ecommerce
   var smUrl = parseUri url
   smUrl.path = "sitemap.xml"
+  smUrl.query = ""
   
   for sitemap in (await fetchSitemap smUrl).querySelectorAll "sitemap":
-    if not sitemap.isNil and sitemap.kind == xnElement:
-      let smSmUrl = sitemap.child("loc").innerText
-      echo smSmUrl
-      # if "products" in smSmUrl:
-      #   for url in (await fetchSitemap smSmUrl).querySelectorAll "url":
-      #     if not url.isNil and url.kind == xnElement:
-      #       echo url
-      #   break
-  
+    let smSmUrl = sitemap.child("loc").innerText
+    if "products" in smSmUrl:
+      for url in (await fetchSitemap smSmUrl).querySelectorAll "url":
+        result.add url.child("loc").innerText
   
 when isMainModule:
-  echo waitFor(extractShopify "https://lojaprecohonesto.com.br/products/camisa-sao-paulo-away-2023")[]
+  echo len waitFor getAllProducts "https://lojaprecohonesto.com.br/products/camisa-sao-paulo-away-2023"