+++ src/shopifyextractor/types.nim
@@ -1,8 +1,4 @@
 type
-  ShopifyEcommerce* = ref object
-    url*: string
-    name*: string
-    products*: seq[string]
   ShopifyProduct* = ref object
     url*: string
     title*: string